Fixing more test mesh failures
git-svn-id: svn+ssh://code.matforge.org/fipy/branches/parallel@3149 d80e17d7-ff13-0410-a124-85740d801063

+++ fipy/variables/meshVariable.py
@@ -106,7 +106,7 @@ class _MeshVariable(Variable):
                 localValue = localValue[..., localIDs]
             globalIDs = numerix.concatenate(comm.allgather(globalIDs))
             
-            globalValue = numerix.empty(localValue.shape[:-1] + (max(globalIDs) + 1,))
+            globalValue = numerix.empty(localValue.shape[:-1] + (max(globalIDs) + 1,), dtype=localValue.dtype)
             globalValue[..., globalIDs] = numerix.concatenate(comm.allgather(localValue), axis=-1)
             
             return globalValue
